properly formatted the code
